Tests/LibWeb/Text/input/css-has-invalidation/baseline-sibling-has.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script src="_helpers.js"></script>
<style>
    .anchor:has(+ .match) { color: red; }
</style>
<div id="a1">
    <div id="a2">
        <span id="anchor" class="anchor">anchor</span>
        <span id="target">target</span>
    </div>
</div>
<script>
    test(() => {
        getComputedStyle(document.getElementById("anchor")).color;
        internals.resetStyleInvalidationCounters();

        // Sibling combinator in :has(). Anchor has sibling-combinator flag,
        // so invalidation must also walk siblings of each visited ancestor.
        document.getElementById("target").classList.add("match");
        getComputedStyle(document.getElementById("anchor")).color;
        printCounters("after add .match on adjacent sibling");
    });
</script>
